+++ src/plugins/waitlistBans.js
@@ -25,9 +25,6 @@ export default function waitlistBansPlugin () {
           timestamp: new Date()
         })
         mp.emit('modWaitlistBan', wrappedBan)
-        // back compat with v1.10.0:
-        // (TODO remove in v2)
-        mp.emit('waitlistBan', wrappedBan)
       })
     })